Move getTrimmedStackTrace() to Throwables
